Create simple LeNet5 model for image training
This is just the backwards steps I've had to take in order to make the original code work.
Several things have changed since this code was originally written. For starters the
original code used Galaxy10 dataset which now uses images 256x256. This may be
implemented later when I'm comfortable with the train and test data from this set.
pytorch is a new addition, as the original code used tensorflow for most of its
model training. The input shapes have been resized to better work with LeNet5 algorithm.

TO DO:Recreate the training/valdiation/testing functions. Save weights, and test on random
sampled data.